+++ tests/testthat/test_growth.R
@@ -4,8 +4,8 @@ data(exampleforestMED)
 data(SpParamsMED)
 data(examplemeteo)
 examplemeteo2 <- examplemeteo
-examplemeteo2$dates <- as.Date(row.names(examplemeteo2))
-row.names(examplemeteo2) <- NULL
+row.names(examplemeteo2) <- as.character(examplemeteo2$dates)
+examplemeteo2$dates <- NULL
 
 control_granier <- defaultControl("Granier")
 control_granier$verbose <- FALSE