+++ NEWS.md
@@ -15,6 +15,7 @@
 - New function 'extract'
 - New infiltration model Green-Ampt (1911)
 - Interception model Liu (2001) accepted as alternative to Gash (1995)
+- Optional "RainfallIntensity" (mm/h) as input in weather
 
 # Version 3.2.0
 - Correction of a bug arisen in old Rcpp versions